+++ peachjam/admin.py
@@ -49,6 +49,7 @@ from peachjam.models import (
     Attorney,
     Author,
     Bench,
+    Bill,
     Book,
     CaseHistory,
     CaseNumber,
@@ -101,6 +102,7 @@ from peachjam.plugins import plugins
 from peachjam.resources import (
     ArticleResource,
     AttorneyResource,
+    BillResource,
     GazetteResource,
     GenericDocumentResource,
     JudgmentResource,
@@ -166,6 +168,9 @@ class EntityProfileInline(GenericStackedInline):
     model = EntityProfile
     extra = 0
     form = EntityProfileForm
+
+
+admin.site.register(PeachJamSettings)
 
 
 class PeachJamSettingsAdmin(admin.ModelAdmin):
@@ -798,6 +803,9 @@ class TaxonomyForm(MoveNodeForm):
         return self.cleaned_data
 
 
+admin.site.register(Taxonomy)
+
+
 class TaxonomyAdmin(TreeAdmin):
     form = movenodeform_factory(Taxonomy, TaxonomyForm)
     readonly_fields = ("slug",)
@@ -823,10 +831,14 @@ class TaxonomyAdmin(TreeAdmin):
         return resp
 
 
+admin.site.register(CoreDocument)
+
+
 class CoreDocumentAdmin(DocumentAdmin):
     pass
 
 
+@admin.register(GenericDocument)
 class GenericDocumentAdmin(ImportExportMixin, DocumentAdmin):
     resource_class = GenericDocumentResource
     fieldsets = copy.deepcopy(DocumentAdmin.fieldsets)
@@ -841,6 +853,9 @@ class GenericDocumentAdmin(ImportExportMixin, DocumentAdmin):
         return qs
 
 
+admin.site.register(Legislation)
+
+
 class LegislationAdmin(ImportExportMixin, DocumentAdmin):
     fieldsets = copy.deepcopy(DocumentAdmin.fieldsets)
     fieldsets[0][1]["fields"].extend(["nature"])
@@ -850,6 +865,11 @@ class LegislationAdmin(ImportExportMixin, DocumentAdmin):
     fieldsets[2][1]["classes"] = ("collapse",)
     fieldsets[4][1]["fields"].extend(["parent_work"])
     readonly_fields = ["parent_work"] + list(DocumentAdmin.readonly_fields)
+
+
+@admin.register(Bill)
+class BillAdmin(ImportExportMixin, DocumentAdmin):
+    resource_class = BillResource
 
 
 class CaseNumberAdmin(admin.StackedInline):
@@ -958,6 +978,9 @@ class JudgmentAdminForm(DocumentForm):
             # re-assigned
             self.instance.serial_number = None
         return super().save(*args, **kwargs)
+
+
+admin.site.register(Judgment)
 
 
 class JudgmentAdmin(ImportExportMixin, DocumentAdmin):
@@ -1562,13 +1585,6 @@ admin.site.register(
         SavedDocument,
     ]
 )
-admin.site.register(PeachJamSettings, PeachJamSettingsAdmin)
-admin.site.register(Taxonomy, TaxonomyAdmin)
-admin.site.register(CoreDocument, CoreDocumentAdmin)
-admin.site.register(GenericDocument, GenericDocumentAdmin)
-admin.site.register(Legislation, LegislationAdmin)
-admin.site.register(Judgment, JudgmentAdmin)
-
 admin.site.unregister(User)
 admin.site.register(User, UserAdminCustom)
