+++ peachjam/templates/peachjam/bill_detail.html
@@ -1,0 +1,1 @@
+{% extends 'peachjam/layouts/document_detail.html' %}
