+++ peachjam/templates/peachjam/bill_list.html
@@ -1,0 +1,8 @@
+{% extends 'peachjam/layouts/document_list.html' %}
+{% load i18n %}
+{% block title %}
+  {% trans 'Bills' %}
+{% endblock %}
+{% block page-title %}
+  <h1 class="my-4">{% trans 'Bills' %}</h1>
+{% endblock %}
